discovery-orchestrator/lambda/requirements.txt
google-cloud-pubsub>=2.19.0
boto3>=1.28.0
